+++ src/pages/constructor/constructor.js
@@ -1,106 +1,15 @@
-import React, { useCallback, useEffect, useRef } from 'react';
+import React, { useEffect, useRef } from 'react';
 import { FormControl, FormHelperText, NativeSelect } from '@material-ui/core';
 import mergeImages from 'merge-images';
-import { useDispatch, useSelector } from 'react-redux';
 import { useStyles } from './constructor.style';
 import { CONSTRUCTOR_TITLES } from '../../translations/constructor.translations';
-import {
-  getConstructorModelById,
-  getModelForConstructor,
-  setModelLoading
-} from '../../redux/constructor/constructor-model/constructor-model.actions';
-import { getConstructorBasic } from '../../redux/constructor/constructor-basic/constructor-basic.actions';
-import { getConstructorPattern } from '../../redux/constructor/constructor-pattern/constructor-pattern.actions';
-import { getConstructorBottom } from '../../redux/constructor/constructor-bottom/constructor-bottom.actions';
-import { getConstructorFrontPocket } from '../../redux/constructor/constructor-front-pocket/constructor-front-pocket.actions';
+import { setModelLoading } from '../../redux/constructor/constructor-model/constructor-model.actions';
 import Loader from '../../components/loader';
-import { selectConstructor } from '../../redux/selectors/multiple.selectors';
+import { useConstructor } from './hooks';
 
 const map = require('lodash/map');
 
 const { MODEL, BASIC, PATTERN, BOTTOM } = CONSTRUCTOR_TITLES[0];
-
-const useConstructor = () => {
-  const dispatch = useDispatch();
-  const {
-    constructorModel,
-    currentModel,
-    basicImage,
-    frontPocketImage,
-    patternImage,
-    bottomImage,
-    modelLoading
-  } = useSelector(selectConstructor);
-
-  const models = constructorModel.modelsForConstructor;
-  const basics = currentModel.constructorBasic;
-  const patterns = currentModel.constructorPattern;
-  const bottoms = currentModel.constructorBottom;
-
-  useEffect(() => {
-    dispatch(getModelForConstructor());
-  }, []);
-
-  useEffect(() => {
-    if (models) {
-      dispatch(getConstructorModelById(models[0]._id));
-    }
-  }, [models]);
-
-  useEffect(() => {
-    if (currentModel) {
-      dispatch(setModelLoading(true));
-      dispatch(getConstructorBasic(currentModel.constructorBasic[0]._id));
-      dispatch(
-        getConstructorFrontPocket(currentModel.constructorFrontPocket[0]._id)
-      );
-      dispatch(getConstructorPattern(currentModel.constructorPattern[0]._id));
-      dispatch(getConstructorBottom(currentModel.constructorBottom[0]._id));
-    }
-  }, [currentModel]);
-
-  const changeModel = useCallback((id) => {
-    dispatch(getConstructorModelById(id));
-  }, []);
-
-  const changeBasic = useCallback((id) => {
-    dispatch(setModelLoading(true));
-    dispatch(getConstructorBasic(id));
-  }, []);
-
-  const changePattern = useCallback((id) => {
-    dispatch(setModelLoading(true));
-    dispatch(getConstructorPattern(id));
-  }, []);
-
-  const changeBottom = useCallback((id) => {
-    dispatch(setModelLoading(true));
-    dispatch(getConstructorBottom(id));
-  }, []);
-
-  return {
-    values: {
-      models,
-      basics,
-      patterns,
-      bottoms,
-      modelLoading
-    },
-    images: {
-      basicImage,
-      frontPocketImage,
-      patternImage,
-      bottomImage
-    },
-    methods: {
-      dispatch,
-      changeModel,
-      changeBasic,
-      changePattern,
-      changeBottom
-    }
-  };
-};
 
 const Constructor = () => {
   const styles = useStyles();
